pyproject.toml
[project]
name = "dockerandmysql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "mysql-connector-python>=9.6.0",
    "ollama>=0.6.1",
]
